+++ resources/views/familia/index.blade.php
@@ -15,9 +15,11 @@
     <br>
       <div class="row">
           <div class="col-md-8">
+          @can('Crear Familia')
         <button id='btnAgregar' onclick="mostrarFormulario()" class="btn btn-success btn-lg">
             Nueva Familia
         </button>
+        @endcan
         <br>
         <br>
               {{ Form::open(['route'=>'familia.store', 'method'=>'POST', 'class'=>'agregar']) }}
@@ -62,11 +64,15 @@
         <td>
           <a class="btn btn-info btn-sm" data-toggle="tooltip" data-placement="top" title="Detalles" href="{{route('familia.show',$value->id)}}">
               <i class="glyphicon glyphicon-list-alt">Detalles</i></a>
+          @can('Editar Familia')
           <a class="btn btn-primary btn-sm" data-toggle="tooltip" data-placement="top" title="Editar" href="{{route('familia.edit',$value->id)}}">
               <i class="glyphicon glyphicon-pencil">Editar</i></a>
+          @endcan
+          @can('Eliminar Familia')
             {!! Form::open(['method' => 'DELETE','route' => ['familia.destroy', $value->id],'style'=>'display:inline', 'class'=>'formulario-eliminar']) !!}
               <button type="submit" data-toggle="tooltip" data-placement="top" title="Eliminar" style="display: inline;" class="btn btn-danger btn-sm">Eliminar<i class="glyphicon glyphicon-trash" ></i></button>
             {!! Form::close() !!}
+          @endcan
         </td>
       </tr>
     @endforeach
