+++ resources/views/users/index.blade.php
@@ -5,10 +5,12 @@
 <div class="row">
     <div class="col-lg-12 margin-tb">
         <div class="pull-left">
-            <h2>Users Management</h2>
+            <h2>Manejo de Usuarios</h2>
         </div>
         <div class="pull-right">
-            <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+        @can('Crear Usuario')
+            <a class="btn btn-success" href="{{ route('users.create') }}"> Crear Usuario</a>
+        @endcan
         </div>
     </div>
 </div>
@@ -24,8 +26,8 @@
 <table class="table table-bordered">
  <tr>
    <th>No</th>
-   <th>Name</th>
-   <th>Email</th>
+   <th>Nombre</th>
+   <th>Correo</th>
    <th>Roles</th>
    <th width="280px">Action</th>
  </tr>
@@ -45,11 +47,15 @@ $i=0;
       @endif
     </td>
     <td>
-       <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-       <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+       <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Detalle</a>
+       @can('Editar Usuario')
+       <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Editar</a>
+       @endcan
+       @can('Eliminar Usuario')
         {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
             {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
         {!! Form::close() !!}
+       @endcan
     </td>
   </tr>
  @endforeach
